Update Toleranzpolitik.md
Have been playing around a bit in Latex, looking at fonts that can be used in an online latex compiler.
Probably nice to be able to use the same document. 
Also looking at textures for the background- slightly stained document look, in black and white, as if scanned much later, would be great.